added lecture 2 on circuits
Learning goals:
What is a circuit, voltage, current, and resistance
What is a resistor and how to use it
What is an LED and how to use it (e.g., polarity, current limiting resistor)
What is Ohm’s Law and why is it useful?
Series vs. parallel resistive circuit configurations
What is a breadboard and how to use it
How to read a datasheet