Create CSE599_L05-PhysicalComputing4-AnalogInput.pdf
Today's Learning Goals
Working with analog input
What is a knob/trim potentiometer and how to use it?
What is a slide potentiometer and how to use it?
What are variable resistive sensors and how to use them?
Understanding the importance of voltage dividers when working
with analog input (and the relevancy of Ohm’s law!)